Fixes Issues Regarding Finding Spider and Swarmer Spawns (#55054)
This PR fixes the issue where people couldn't teleport to spider egg clusters using the notification.

It also adds both the swarmer beacon and spider egg clusters to the ghost orbit menu. This should make it easier to find instances of both in cases where a player might miss the initial notification.