Tweaks to temperature code (#55216)
This is a small tweak that makes skin temp raise up faster when cold in response to ice moon being harder.
This also makes a change to mobs in statis beds no long hold bed temp in some quantum lock but stops the natural stabilization. This means that is you put a frozen/superheated mob on the floor or a statis bed they will balance tot the room temp over time. The mob is not physically separate from the room in a statis bed and all life functions are still suspended.