Sets default food size to small. (#55210)
## About The Pull Request

Fixes an oversight done during the food refactor that makes all food items normal sized instead of small.
Instead of adding weight class to a bunch of individual items [(like this PR)](https://github.com/tgstation/tgstation/pull/55174) it just makes all food default to small unless tagged to be bigger.

## Why It's Good For The Game

Being able to put 7 items on a tray instead of 4 is good, and if a food item is "too powerful", that individual item (or subtype, such as soups) can be given a larger weight class.

Fixes #54818