Fix roulette wheel tgui error (#55376)
Fixes #54394

Once upon a time, tgui probably sent numbers to DM as a string. Now they're sent as actual numbers. Progress!

Roulette wheel seems to expect strings DM-side (and has valid text inputs from tgui) and the JS code also expected strings in `data`.

If you select any of the single bets, it sends a non-string type back to DM which then feeds back as a non-string type to JS, it then attempts to call `startsWith` on this non-string type and triggers the script error.

I've opted to fix this tgui-side by making sure it sends only strings back to the game, which allowed me to use the roulette wheel again, bet on singles without crashing and indeed in testing I could successfully win on a single as well. So it appears to restore full functionality.