tgui: Cyborg Boot Debug (#55584)
This PR refactors Cyborg Boot Debug and brings it's UI from html based one to tgui with some improvements and cleanup. I also took a liberty to add tooltip to each setting to shed some light on its function, as it could be quite unclear what exactly it meant.
Conflicts:
	tgui/public/tgui-common.chunk.js
	tgui/public/tgui-panel.bundle.js
	tgui/public/tgui.bundle.js