Better saving of chat logs (#55898)
Makes it so that it'll only save the messages that are visible in the current tab. ie, the logs will respect the types of messages in the current tab.

Because when you just want admin messages or combat logs, it's annoying to have to sort through literally everything else to get them.
Conflicts:
	tgui/public/tgui-panel.bundle.js