tgui: Civilian Bounties, choice of three (#55762)
The civilian bounty console now offers a choice of 3 different bounties when getting a new civilian bounty choice.
The first two of these options will be different bounty types from within your jobs pool, while the third will always be a generic bounty (AKA from the pool that assistants and various misc. jobs choose from).
This gives the player a bit more choice than they had before in regards to doing a task related to their job within the existing bounty system until we add a more concrete method for some of the more poorly aligned bounty pools to be brought in-line with existing gameplay.
Also it offers a way to make economy impact more interactive with the round.

Co-authored-by: fuck <minecraftor5088@gmail.com>
Conflicts:
	tgui/public/tgui.bundle.js