Third Rate Duelist With A Fourth Rate Deck: The TG TCG, Part 2 (#54659)
Finally adds the ruleset to the TGC, allowing players to play the game (in a relatively provisional state).
Cards now have rules and keywords attached, as well as a proper system for displaying information in the sidebar (all the credit for this has to go to @ArcaneMusic, who also contributed a huge amount of the ruleset for the cards).
The full ruleset and keywords can be found here: https://tgstation13.org/wiki/Tactical_Game_Cards