Fix chemical grafts (#56000)
Fixes chemical grafts in botany so their seeds actually gain the reagents they offer.

For those who didn't know, Spaceman's Trumpet is the only plant (for now) that allows you you graft it and get a chemical instead of a trait (Polypyrylium Oligomers). Only this graft doesn't work. It shows up in the plant, but on harvest there is none of that chemical inside of it. Now it works.