Move psychologist immunity from mind to skillchip (#55813)
Skillchips are neat, and this is the sort of "job-specific skill" that can remain job specific, but still gainable in the shift through in-universe means.

The logic behind the skillchip is that the supermatter is psychically sensitive, hence the hallucinations and headaches from looking at it and projecting the "correct mental aura" will help it calm down. But that is lore following mechanics.

Refactors skillchips to be able to automatically apply more than one trait.
Conflicts:
	code/__DEFINES/traits.dm
	code/modules/library/skill_learning/skillchip.dm
	code/modules/power/supermatter/supermatter.dm