Fixes thermomachine power bug (#56261)
fixes #56257 .... Gilkher when he made thermomachines he didnt add in not a single check to make sure the machine actually had power to use, so was able to be used in space, with no power, so long as the machine was on. I simply added the proc that processed its gas contents behind a simple check to ensure it has power, isnt broken, and the maint cover is closed. if one of those is true, it will turn the machine back off and set power usage to 0

Before the fix:
https://user-images.githubusercontent.com/40489693/104986520-28165d00-59e1-11eb-9402-a883baa61dea.mp4

After the fix:
https://i.gyazo.com/3e3392095797a8ffc78178504117be8e.mp4
Conflicts:
	code/modules/atmospherics/machinery/components/unary_devices/thermomachine.dm