Replaces intents with combat mode (#56601)
About The Pull Request

This PR removes intents and replaces them with a combat mode. An explanation of what this means can be found below
Major changes:

    Disarm and Grab intents have been removed.
    Harm/Help is now combat mode, toggled by F or 4 by default
    The context/verb/popup menu now only works when you do shift+right-click
    Right click is now disarm, both in and out of combat mode.
    Grabbing is now on ctrl-click.
    If you're in combat mode, and are currently grabbing/pulling someone, and ctrl-click somewhere else, it will not release the grab (To prevent misclicks)

Minor interaction changes:

Right click to dissasemble tables, racks, filing cabinets (When holding the right tool to do so)
Left click to stunbaton, right click to harmbaton
Right click to tip cows
Right click to malpractice surgery
Right click to hold people at gunpoint (if youre holding a gun)
Why It's Good For The Game

Intents heavily cripple both the code and the UI design of interactions. While I understand that a lot of people will dislike this PR as they are used to intents, they are one of our weakest links in terms of explaining to players how to do specific things, and require a lot more keypresses to do compared to this.

As an example, martial arts can now be done without having to juggle 1 2 3 and 4 to switch intents quickly.

As some of you who saw the first combat mode PR, the context menu used to be disabled in combat mode. In this version it is instead on shift-right click ensuring that you can always use it in the same way.

In this version, combat mode also no longer prevents you from attacking with items when you would so before, as this was something that was commonly complained about.

The full intention of this shift in control scheme is that right click will become "secondary interaction" for items, which prevents some of the awkward juggling we have now with item modes etcetera.
Changelog

cl Qustinnus
add: Intents have been replaced with a combat mode. For more info find the PR here: #56601
/cl
Conflicts:
	code/__DEFINES/monkeys.dm
	code/_onclick/other_mobs.dm
	code/datums/ai/monkey/monkey_behaviors.dm
	code/datums/ai/monkey/monkey_controller.dm
	code/datums/components/material_container.dm
	code/datums/components/riding/riding_mob.dm
	code/datums/martial/cqc.dm
	code/datums/martial/psychotic_brawl.dm
	code/game/objects/structures/window.dm
	code/modules/antagonists/eldritch_cult/knowledge/rust_lore.dm
	code/modules/aquarium/aquarium.dm
	code/modules/mob/living/carbon/carbon_defense.dm
	code/modules/mob/living/carbon/human/human.dm
	code/modules/mob/living/carbon/human/human_defense.dm
	code/modules/mob/living/carbon/human/species.dm
	code/modules/mob/living/carbon/human/species_types/golems.dm
	code/modules/mob/living/carbon/human/species_types/monkeys.dm
	code/modules/mob/living/living.dm
	code/modules/mob/living/living_defense.dm
	code/modules/mob/living/simple_animal/animal_defense.dm
	code/modules/mob/living/simple_animal/hostile/giant_spider.dm
	code/modules/mob/mob_helpers.dm
	code/modules/modular_computers/computers/machinery/modular_computer.dm
	code/modules/surgery/tools.dm
	code/modules/unit_tests/metabolizing.dm
	icons/hud/screen_gen.dmi