Refactor ventcrawling flag into traits (#56620)
Instead of using var/ventcrawling on `/mob/living`, it is now instead
two traits. It functions in exactly the same way.

This now ensures that manipulation of ventcrawling will not clash with
any other manipulation, such as a proposed genetics ability to give
people ventcrawling, versus abductor organs.

---

Intended to be a pure refactor, no functionality should change.
Conflicts:
	code/datums/diseases/transformation.dm
	code/modules/mob/living/carbon/alien/alien.dm
	code/modules/mob/living/carbon/human/species_types/monkeys.dm
	code/modules/mob/living/simple_animal/hostile/cockroach.dm
	code/modules/mob/living/simple_animal/hostile/mining_mobs/gutlunch.dm