Makes cryo handle reagent transfer better. (#56755)
Cryogenics has a reagent_transfer variable that is set to throttle the
reagent injection. Upgrading bins increase efficiency and transfer
amount, thus making the reagent_transfer threshold larger, making it
take longer between each injections.

However, this variable is stored
in the object itself and not on the process proc, presumably to prevent
exploits such as rapidly turning the machine on and off. This has a
really bad consequence that makes cryo really clunky to use, since you
may need to wait for the reagent_transfer variable to hit threshold and
then reset again for a new patient before any cryoxadone can be
injected. Not to mention how upgrading the bins somehow timegaps it
even further.

This PR seeks to change the whole system to transfer reagents in a
smaller and more continuous manner.