Add blended dust warning to grinder beakers (#56759)
* Add blended dust warning to grinder beakers

:cl: coiax
tweak: All beakers, bottles and containers that have been used in the
All-in-One Grinder will have a warning not to inhale the dust on
examine.
/:cl:

The initial spawned beakers in the grinder already had this description,
and now it is applied to any beakers that have been used for grinding
purposes (and only grinding purposes).

Co-authored-by: cacogen <25089914+cacogen@users.noreply.github.com>
Conflicts:
	code/__DEFINES/traits.dm