Fix dynamic parameters ignoring configuration (#56778)
Dynamic was ignoring the parameters in dynamic.json, because they were being ran in can_start, which runs before pre_setup. This has been moved to its own proc which is correctly called after configuration.

Let this be a lesson in why you should not be putting side effects in functions that don't imply they are impure.
Conflicts:
	code/game/gamemodes/dynamic/dynamic.dm