Fixes bullying socially anxious dead people by blowing kisses at their corpse (#56732)
At the end of #56698 I added interactions for people with the social anxiety quirk so they sometimes get flustered when hit by a kiss, but I realized precisely 0.2 seconds too late that I didn't actually check if they were alive or conscious before doing so.

This PR fixes this so those reactions will only happen if they're conscious or in soft crit. I also changed the description of the kiss of death lipstick, cause I forgot to put one
Conflicts:
	code/game/objects/items/cosmetics.dm
	code/game/objects/items/hand_items.dm
	code/modules/mob/living/emote.dm